# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sistema de Gestión - Reparadora Fercho",
    "description": "API completa para el sistema de gestión de zapatería 'Reparadora Fercho'. Incluye gestión de clientes, pedidos y cobros con validaciones y ejemplos detallados.",
    "version": "2.0.0",
    "contact": {
      "name": "Reparadora Fercho",
      "email": "admin@reparadorafercho.com",
      "url": "http://localhost:8080"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desarrollo local"
    }
  ],
  "paths": {
...
    "/api/pedidos": {
      "get": {
        "summary": "Obtener todos los pedidos",
        "description": "Retorna la lista de todos los pedidos con información del cliente asociado, ordenados por fecha de creación (más recientes primero).",
        "tags": ["Pedidos"],
        "responses": {
          "200": {
            "description": "Lista de pedidos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pedido"
                      },
                      "example": [
                        {
                          "id": 1,
                          "idCliente": 1,
                          "descripcion": "Reparación de suela",
                          "estado": "Pendiente",
                          "fecha_creacion": "2025-08-08T03:00:00.000Z",
                          "nombreCliente": "Juan Pérez"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nuevo pedido",
        "description": "Crea un nuevo pedido de reparación en el sistema. El cliente debe existir y la descripción es obligatoria.",
        "tags": ["Pedidos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["idCliente", "descripcion"],
                "properties": {
                  "idCliente": {
                    "type": "integer",
                    "description": "ID del cliente que solicita el pedido",
                    "minimum": 1,
                    "example": 1
                  },
                  "descripcion": {
                    "type": "string",
                    "description": "Descripción detallada del trabajo a realizar",
                    "minLength": 10,
                    "maxLength": 500,
                    "example": "Reparación de suela de zapato deportivo, cambio de tacón y ajuste de horma"
                  },
                  "estado": {
                    "type": "string",
                    "description": "Estado actual del pedido",
                    "enum": ["Pendiente", "Entregado"],
                    "default": "Pendiente",
                    "example": "Pendiente"
                  }
                }
              },
              "examples": {
                "pedido_basico": {
                  "summary": "Pedido básico",
                  "value": {
                    "idCliente": 1,
                    "descripcion": "Reparación de suela de zapato",
                    "estado": "Pendiente"
                  }
                },
                "pedido_detallado": {
                  "summary": "Pedido con descripción detallada",
                  "value": {
                    "idCliente": 2,
                    "descripcion": "Cambio completo de suela, reparación de costuras y limpieza general del calzado",
                    "estado": "Pendiente"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Pedido creado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pedido"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "campos_faltantes": {
                    "summary": "Campos obligatorios faltantes",
                    "value": {
                      "success": false,
                      "message": "ID del cliente y descripción son obligatorios"
                    }
                  },
                  "cliente_no_existe": {
                    "summary": "Cliente no encontrado",
                    "value": {
                      "success": false,
                      "message": "Cliente no encontrado"
                    }
                  },
                  "estado_invalido": {
                    "summary": "Estado inválido",
                    "value": {
                      "success": false,
                      "message": "Estado debe ser \"Pendiente\" o \"Entregado\""
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pedido": {
        "type": "object",
        "description": "Información de un pedido de reparación",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único del pedido",
            "example": 1
          },
          "idCliente": {
            "type": "integer",
            "description": "ID del cliente que solicitó el pedido",
            "example": 1
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción del trabajo a realizar",
            "example": "Reparación de suela de zapato"
          },
          "estado": {
            "type": "string",
            "description": "Estado actual del pedido",
            "enum": ["Pendiente", "Entregado"],
            "example": "Pendiente"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de creación del pedido",
            "example": "2025-08-08T03:00:00.000Z"
          },
          "nombreCliente": {
            "type": "string",
            "description": "Nombre del cliente asociado al pedido",
            "example": "Juan Pérez"
          }
        },
        "required": ["id", "idCliente", "descripcion", "estado", "fecha_creacion", "nombreCliente"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Respuesta de error estándar",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del error",
            "example": "Error al procesar la solicitud"
          },
          "error": {
            "type": "object",
            "description": "Detalles adicionales del error (solo en desarrollo)",
            "additionalProperties": true
          }
        },
        "required": ["success", "message"]
      }
    }
  },
...
} 